requirement.txt
pyenchant
nltk
textmining
scipy
numpy
scikit-learn
pandas
flask
gensim
